Setup tests with mocha and chai
