+++ lab-giodamelio/lib/server.js
@@ -1,6 +1,11 @@
 'use strict';
-const createServer = require('./index');
+const net = require('net');
 
-const server = createServer();
-server.listen(3000);
-console.log('Chat server listening at tcp://localhost:3000');
+module.exports = function createServer() {
+  const server = net.createServer((socket) => {
+    socket.write('Hello World');
+    socket.end();
+  });
+
+  return server;
+};
