+++ lab-giodamelio/package.json
@@ -4,14 +4,16 @@
   "description": "A simple tcp chat server",
   "main": "lib/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "mocha test/",
     "lint": "eslint lib/ test/"
   },
   "author": "Gio d'Amelio",
   "license": "MIT",
   "devDependencies": {
+    "chai": "^3.5.0",
     "eslint": "^3.1.1",
     "eslint-config-airbnb-base": "^4.0.2",
-    "eslint-plugin-import": "^1.11.0"
+    "eslint-plugin-import": "^1.11.0",
+    "mocha": "^2.5.3"
   }
 }
