+++ lab-giodamelio/test/.eslintrc
@@ -1,0 +1,10 @@
+{
+  "rules": {
+    "no-unused-expressions": 0,
+    "import/no-extraneous-dependencies": [
+      "error", {
+        "devDependencies": true
+      }
+    ]
+  }
+}
